Major api refactor.

+++ src/main/java/org/sgrewritten/stargate/economy/EconomyAPI.java
@@ -14,7 +14,7 @@ public interface EconomyAPI {
      * @param amount <p>The amount required</p>
      * @return <p>True if the player has the required amount</p>
      */
-    boolean has(OfflinePlayer target, int amount);
+    boolean has(OfflinePlayer target, double amount);
 
     /**
      * Charges a player if possible
@@ -23,7 +23,7 @@ public interface EconomyAPI {
      * @param amount        <p>The amount the player should be charged</p>
      * @return <p>True if the payment was fulfilled</p>
      */
-    boolean chargePlayer(OfflinePlayer offlinePlayer, int amount);
+    boolean chargePlayer(OfflinePlayer offlinePlayer, double amount);
 
     /**
      * Deposits money in a player's account if possible
@@ -32,6 +32,6 @@ public interface EconomyAPI {
      * @param amount <p>The amount to receive</p>
      * @return <p>True if the payment was fulfilled</p>
      */
-    boolean depositPlayer(OfflinePlayer player, int amount);
+    boolean depositPlayer(OfflinePlayer player, double amount);
 
 }
